+++ test/dummy/app/models/passport.rb
@@ -8,7 +8,7 @@ class Passport < ActiveRecord::Base
   aggregate_attribute :city,             :string,   required: true
   aggregate_attribute :state,            :string,   required: true
   aggregate_attribute :birthdate,        :datetime, required: true
-  aggregate_attribute :height,           :decimal
+  aggregate_attribute :height,           :decimal, track_all_values: true
   aggregate_attribute :weight,           :decimal, default: 100
   aggregate_attribute :photo,            "PassportPhoto"
   aggregate_has_many  :foreign_visits,   "ForeignVisit"